Refactored load-test to use one slack token
