+++ tools/load-test/Dockerfile
@@ -1,7 +1,7 @@
 FROM alpine:3.8
 LABEL source="git@github.com:kyma-project/kyma.git"
 RUN apk --no-cache upgrade && apk --no-cache add curl
-RUN curl -Lo /usr/bin/kubectl https://storage.googleapis.com/kubernetes-release/release/v1.10.0/bin/linux/amd64/kubectl && chmod +x /usr/bin/kubectl
+RUN curl -Lo /usr/bin/kubectl https://storage.googleapis.com/kubernetes-release/release/v1.11.0/bin/linux/amd64/kubectl && chmod +x /usr/bin/kubectl
 # To automatically get the latest version:
 #RUN curl -Lo /usr/bin/kubectl https://storage.googleapis.com/kubernetes-release/release/v/bin/linux/amd64/kubectl && chmod +x /usr/bin/kubectl
 RUN mkdir -p /root/.kube && touch /root/.kube/config
